Introduce a read-only "maintenance mode" that can be toggled on or off via an API endpoint
While in maintenance mode, API operations that mutate data in the
datastore will fail with 503 Service Unavailable. This provides a way to
ensure that new references to blobs aren't added while a garbage
collection pass runs.

Maintenance mode is switched on of off by POSTing to /admin/maintenance/on
or /admin/maintenance/off. This endpoint is only available to
authenticated users that have full admin access to the resource with
type "registry", and name "administration".

WIP, not ready to merge.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>